modify view_package job to handle updated asset file paths instead of directory, also allowing for two different config file naming conventions

+++ app/jobs/get_view_packages_from_github.rb
@@ -19,27 +19,40 @@ class GetViewPackagesFromGithub
     g.pull()
     g.checkout('feature/config-file') #for now TODO change this to develop or master when it is merged in
     FileUtils.cd(Rails.root.join("public/view_packages"))
+    system("git pull")
     system("git submodule update --init") #update or initialize any submodules, like tapas-generic
     directories = Dir.glob('*').select {|f| File.directory? f}
-    view_packages = available_view_packages
+    view_packages = ViewPackage.where("").pluck(:machine_name).to_a
     directories.each do |dir_name|
       dir = dir_name.sub("-","_")
       if view_packages.include?(dir)
         view = ViewPackage.where(machine_name: dir).first
       else
-        view = ViewPackage.new(:machine_name => dir)
+        view = ViewPackage.create(:machine_name => dir)
       end
       FileUtils.cd(Rails.root.join("public/view_packages/#{dir_name}"))
       # look for config file
       puts view.inspect
-      if File.exist?("PKG-CONFIG.xml")
+      if File.exist?("PKG-CONFIG.xml") || File.exist?("CONFIG.xml")
         puts "config file exists"
-        doc = File.open("PKG-CONFIG.xml") { |f| Nokogiri::XML(f) }
+        if File.exist?("PKG-CONFIG.xml")
+          doc = File.open("PKG-CONFIG.xml") { |f| Nokogiri::XML(f) }
+        elsif File.exist?("CONFIG.xml")
+          doc = File.open("CONFIG.xml") { |f| Nokogiri::XML(f) }
+        end
         puts doc
         view.human_name = doc.css("view_package human_name").text
         view.description = doc.css("view_package description").text
-        view.css_dir = doc.css("view_package css_dir").text
-        view.js_dir = doc.css("view_package js_dir").text
+        css_files = []
+        doc.css("view_package css_files css_file").each do |file|
+          css_files << file.text
+        end
+        view.css_files = css_files
+        js_files = []
+        doc.css("view_package js_files js_file").each do |file|
+          js_files << file.text
+        end
+        view.js_files = js_files
         file_types = []
         doc.css("view_package file_types file_type").each do |type|
           file_types << type.text
